Use ulong for server id

+++ RaftNET.Tests/ReplicationTests/Connection.cs
@@ -1,3 +1,3 @@
 ﻿namespace RaftNET.Tests.ReplicationTests;
 
-public record Connection(int from, int to);
+public record Connection(ulong from, ulong to);
