+++ RaftNET.Tests/ReplicationTests/Entries.cs
@@ -1,3 +1,3 @@
 ﻿namespace RaftNET.Tests.ReplicationTests;
 
-public record Entries(int N, int? Server = null, bool Concurrent = false);
+public record Entries(ulong N, ulong? Server = null, bool Concurrent = false);
