+++ RaftNET.Tests/ReplicationTests/SnapshotValue.cs
@@ -2,5 +2,5 @@
 
 public record SnapshotValue {
     public HasherInt Hasher = new(false);
-    public int Idx;
+    public ulong Idx;
 }
